requirements.txt
# Agent GUI Autonome - Dependencies
pyautogui>=0.9.54
Pillow>=10.0.0
requests>=2.31.0
pyperclip>=1.8.0
